Add support for sub routers with parameterized routes
Previously, when using a sub router on a parameterized route like
`app.use('/foo/:param', router)`, `findRouteName` would not correctly
resolve the route parameter for the parent route. This happened because
`req.baseUrl` contained the url with the value, not the parameter. To
solve this, this PR adds code to translate from the url with the value,
back to the parameterized version (i.e. `/foo/value -> /foo/:param`).